settings and saving
more work on the settings and saving specific pages. Gallery page is also working in a basic form, just need to hook it up to the main page to know which page is currently being shown. also need to refactor the pages and viewmodels to pull out the common stuff