Default Page
Default Page feature should now work. If user specified a page to be the default it will load first, otherwise the app opens to the Instagram homepage. Added a BoolToColorConverter so the Gallery list can highlight the default page. Also added a check to hopefully make sure only one page is default at a time.