Cache bugfix
Fixed an issue where the WebView's cookie cache would cause a login dialog to block user interaction after a while of using the app. Added a new WebService that allows for clearing the cookie cache which can be called from a new button in Settings and is also called on App Sleep. Added a new subclass of WebView to support custom renderers and a new Android custom renderer.